+++ react/ProductSearchContextProvider.js
@@ -9,6 +9,10 @@ import { SearchQueryContext } from './constants/searchContext'
 class ProductSearchContextProvider extends Component {
   static propTypes = searchContextPropTypes
 
+  stripCategory(category) {
+    return category.replace(/^\/|\/$/g, '')
+  }
+
   getData = () => {
     const { searchQuery } = this.props
 
@@ -25,7 +29,7 @@ class ProductSearchContextProvider extends Component {
           name: product.productName,
           list: 'Search Results',
           brand: product.brand,
-          category: path(['categories', '0'], product),
+          category: this.stripCategory(path(['categories', '0'], product)),
           position: index + 1,
         })),
       },