+++ react/StoreContextProvider.js
@@ -2,11 +2,13 @@ import React, { Component } from 'react'
 import { Helmet } from 'render'
 import PropTypes from 'prop-types'
 
-import { gtmScript, gtmFrame} from './scripts/gtm'
+import { OrderFormProvider } from './OrderFormContext'
 import { DataLayerProvider } from './components/withDataLayer'
+import { gtmScript, gtmFrame } from './scripts/gtm'
 
 const APP_LOCATOR = 'vtex.store'
-
+const CONTENT_TYPE = 'text/html;charset=utf-8'
+const META_ROBOTS = 'index, follow'
 class StoreContextProvider extends Component {
   static propTypes = {
     children: PropTypes.element,
@@ -17,14 +19,22 @@ class StoreContextProvider extends Component {
   }
 
   render() {
-    const settings = this.context.getSettings(APP_LOCATOR) || {}
     window.dataLayer = window.dataLayer || []
-    const {gtmId} = settings
+    const { country, locale, currency } = global.__RUNTIME__.culture
+    const settings = this.context.getSettings(APP_LOCATOR) || {}
+    console.log(settings)
+    const {
+      gtmId,
+      titleTag,
+      metaTagDescription,
+      metaTagKeywords,
+      storeName,
+    } = settings
     const scripts = gtmId ? [{
       'type': 'application/javascript',
       'innerHTML': gtmScript(gtmId),
     }] : []
-    const noscripts = gtmId ? [{id: "gtm_frame", innerHTML: gtmFrame(gtmId)}] : []
+    const noscripts = gtmId ? [{ id: 'gtm_frame', innerHTML: gtmFrame(gtmId) }] : []
     return (
       <DataLayerProvider
         value={{
@@ -32,8 +42,22 @@ class StoreContextProvider extends Component {
           set: this.pushToDataLayer,
         }}
       >
-      <Helmet script={scripts} noscript={noscripts} />
-      <div className="vtex-store__template">{this.props.children}</div>
+        <Helmet script={scripts} noscript={noscripts} />
+        <Helmet>
+          <title>{titleTag}</title>
+          <meta name="description" content={metaTagDescription} />
+          <meta name="keywords" content={metaTagKeywords} />
+          <meta name="copyright" content={storeName} />
+          <meta name="author" content={storeName} />
+          <meta name="country" content={country} />
+          <meta name="language" content={locale} />
+          <meta name="currency" content={currency} />
+          <meta name="robots" content={META_ROBOTS} />
+          <meta httpEquiv="Content-Type" content={CONTENT_TYPE} />
+        </Helmet>
+        <OrderFormProvider>
+          <div className="vtex-store__template">{this.props.children}</div>
+        </OrderFormProvider>
       </DataLayerProvider>
     )
   }